requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-cors-headers==4.3.1
python-dotenv==1.0.1

# Database
psycopg2>=2.9.10

# Celery
celery==5.3.6
redis==5.0.1

# Email
sendgrid==6.11.0

# Development
django-debug-toolbar==4.3.0
pytest-django==4.7.0
pytest-cov==4.1.0

# Documentation
markdown==3.5.2
Pygments==2.17.2
